requirements.txt
accelerate==0.28.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
bitsandbytes==0.39.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
crcmod==1.7
cryptography==45.0.5
cycler==0.12.1
datasets==2.18.0
decord==0.6.0
deepspeed==0.14.0
dill==0.3.8
distro==1.9.0
docstring-parser==0.16
dotenv==0.9.9
einops==0.8.1
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.110.1
ffmpy==0.5.0
filelock==3.13.4
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.2.0
fuzzywuzzy==0.18.0
gradio==4.26.0
gradio-client==0.15.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib-resources==6.5.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.31.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
openai==1.71.0
openpyxl==3.1.5
orjson==3.10.15
oss2==2.19.1
packaging==24.2
pandas==2.2.3
peft==0.10.0
pillow==10.4.0
propcache==0.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.6.4
pydantic-core==2.16.3
pydub==0.25.1
pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
qwen-vl-utils==0.0.10
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rouge-chinese==1.0.3
rpds-py==0.23.1
ruff==0.11.2
safetensors==0.4.2
semantic-version==2.10.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
sse-starlette==2.1.0
starlette==0.37.2
sympy==1.13.3
termcolor==2.4.0
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.2
transformers==4.49.0
triton==3.0.0
trl==0.8.1
typeguard==4.4.2
typer==0.15.2
typing-extensions==4.11.0
tyro==0.9.17
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.29.0
websockets==11.0.3
xxhash==3.5.0
yarl==1.18.3
flash-attn @ file:///home/dengchao/Downloads/flash_attn-2.6.2+cu123torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl

